moved component inside blog folder
